WrongTest

+++ src/HandShake.cs
@@ -66,7 +66,7 @@ namespace codecrafters_bittorrent.src
             Dictionary<string, object> extensionMsg = new Dictionary<string, object>();
             Dictionary<string, object> extensions = new Dictionary<string, object>();
 
-            extensions.Add("ut_metadata", 222);
+            extensions.Add("ut_metadata", (byte)222);
             extensionMsg.Add("m", extensions);
 
             Console.WriteLine(JsonSerializer.Serialize(extensionMsg));
